update
added some javascript to change the "los angeles" to whatever the user inputs in the search bar. Also, when ℃ is click on it will convert(hard code) 83℉ to 28.3℃. Update the date/time portion to include the current day of the week.
I added more styling to make it look cute.